pyproject.toml
[project]
name = "research-article-writer"
version = "0.1.0"
description = "Agentic AI system for automated systematic literature reviews"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "prisma-flow-diagram>=0.1.0",
    "colrev>=0.15.0",  # Required dependency for prisma-flow-diagram
    "docker>=7.0.0",  # Required dependency for colrev
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "networkx>=3.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-genai>=0.1.0",
    "perplexityai>=0.1.0",
    "exa_py>=0.1.0",
    "tavily-python>=0.1.0",
    "langchain>=0.1.0",
    "langgraph>=0.2.0",
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "mdutils>=1.5.1",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "arxiv>=2.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "rich>=13.0.0",
    "colorama>=0.4.6",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
